% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-ants-preprocess.R
\name{ants_preprocessing}
\alias{ants_preprocessing}
\title{Process 'T1' weighted 'MRI' using \code{'ANTs'}}
\usage{
ants_preprocessing(
  work_path,
  image_path,
  resample = FALSE,
  verbose = TRUE,
  template_subject = raveio_getopt("threeBrain_template_subject")
)
}
\arguments{
\item{work_path}{working directory, all intermediate images will be stored
here}

\item{image_path}{input image path}

\item{resample}{whether to resample the input image before processing}

\item{verbose}{whether to verbose the processing details}

\item{template_subject}{template mapping, default is derived from
\code{\link{raveio_getopt}}}
}
\value{
Nothing. All images are saved to \code{work_path}
}
\description{
This function is soft-deprecated. Use \code{\link{yael_preprocess}} instead.
}
