% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-subject.R
\name{as_rave_subject}
\alias{as_rave_subject}
\title{Get \code{\link{RAVESubject}} instance from character}
\usage{
as_rave_subject(subject_id, strict = TRUE)
}
\arguments{
\item{subject_id}{character in format \code{"project/subject"}}

\item{strict}{whether to check if subject directories exist or not}
}
\value{
\code{\link{RAVESubject}} instance
}
\description{
Get \code{\link{RAVESubject}} instance from character
}
\seealso{
\code{\link{RAVESubject}}
}
