% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{ask_question}
\alias{ask_question}
\title{Ask questions via 'RStudio' or console}
\usage{
ask_question(title, message, ok = "Yes", cancel = "No", use_console = FALSE)
}
\arguments{
\item{title, message}{title and message of the question}

\item{ok, cancel}{the button label to be displayed}

\item{use_console}{whether to force use console to ask questions}
}
\value{
A logical value.
}
\description{
Use 'RStudio' dialogue to ask question and collect
answers interactively. If 'RStudio' is not running, a console version
will be raised.
}
