% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{cache_path}
\alias{cache_path}
\alias{cache_root}
\alias{clear_cached_files}
\title{Manipulate cached data on the file systems}
\usage{
cache_root(check = FALSE)

clear_cached_files(quiet = FALSE)
}
\arguments{
\item{check}{whether to ensure the cache root path}

\item{quiet}{whether to suppress the message}
}
\value{
\code{cache_root} returns the root path that stores the 'RAVE'
cache data; \code{clear_cached_files} returns nothing
}
\description{
Manipulate cached data on the file systems
}
\examples{

cache_root()

}
