% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-aaa.R, R/cmd-dcm2niix.R, R/cmd-flirt.R,
%   R/cmd-recon-all.R
\name{cmd-external}
\alias{cmd-external}
\alias{cmd_run_dcm2niix}
\alias{cmd_run_flirt}
\alias{cmd_run_recon_all}
\title{External shell commands for 'RAVE'}
\usage{
cmd_run_dcm2niix(
  subject,
  src_path,
  type = c("MRI", "CT"),
  merge = c("Auto", "No", "Yes"),
  float = c("Yes", "No"),
  crop = c("No", "Yes", "Ignore"),
  overwrite = FALSE,
  command_path = NULL,
  dry_run = FALSE,
  verbose = dry_run
)

cmd_run_flirt(
  subject,
  mri_path,
  ct_path,
  overwrite = FALSE,
  command_path = NULL,
  dry_run = FALSE,
  verbose = dry_run
)

cmd_run_recon_all(
  subject,
  mri_path,
  args = c("-all", "-autorecon1", "-autorecon2", "-autorecon3", "-autorecon2-cp",
    "-autorecon2-wm", "-autorecon2-pial"),
  work_path = NULL,
  overwrite = FALSE,
  command_path = NULL,
  dry_run = FALSE,
  verbose = dry_run
)
}
\description{
External shell commands for 'RAVE'
}
