% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-knitr.R
\name{configure_knitr}
\alias{configure_knitr}
\title{Configure \code{'rmarkdown'} files to build 'RAVE' pipelines}
\usage{
configure_knitr(languages = c("R", "python"))
}
\arguments{
\item{languages}{one or more programming languages to support; options are
\code{'R'} and \code{'python'}}
}
\value{
A function that is supposed to be called later that builds the
pipeline scripts
}
\description{
Allows building 'RAVE' pipelines from \code{'rmarkdown'} files.
Please use it in \code{'rmarkdown'} scripts only. Use
\code{\link{pipeline_create_template}} to create an example.
}
