% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-BlackrockFile.R
\name{convert_blackrock}
\alias{convert_blackrock}
\title{Convert 'BlackRock' 'NEV/NSx' files}
\usage{
convert_blackrock(
  file,
  block = NULL,
  to = NULL,
  comments = TRUE,
  format = c("mat", "hdf5")
)
}
\arguments{
\item{file}{path to any 'NEV/NSx' file}

\item{block}{the block name, default is file name}

\item{to}{save to path, must be a directory; default is under the file path}

\item{comments}{whether to extract comment section as epoch; default is true}

\item{format}{output format, choices are \code{'mat'} or \code{'hdf5'}}
}
\value{
The results will be stored in directory specified by \code{to}. The
function itself returned nothing.
}
\description{
Convert 'BlackRock' 'NEV/NSx' files
}
