% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rave-meta.R
\name{convert_electrode_table_to_bids}
\alias{convert_electrode_table_to_bids}
\title{Convert electrode table}
\usage{
convert_electrode_table_to_bids(
  subject,
  space = c("ScanRAS", "MNI305", "fsnative")
)
}
\arguments{
\item{subject}{'RAVE' subject}

\item{space}{suggested coordinate space, notice this argument might not be
supported when \code{'FreeSurfer'} reconstruction is missing.}
}
\value{
A list of table in data frame and a list of meta information
}
\description{
Convert electrode table
}
