% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{dir_create}
\alias{dir_create}
\title{Force creating directory with checks}
\usage{
dir_create(x, showWarnings = FALSE, recursive = TRUE, check = TRUE, ...)
}
\arguments{
\item{x}{path to create}

\item{showWarnings, recursive, ...}{passed to \code{\link{dir.create}}}

\item{check}{whether to check the directory after creation}
}
\value{
Normalized path
}
\description{
Force creating directory with checks
}
