% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-project.R
\name{get_projects}
\alias{get_projects}
\title{Get all possible projects in 'RAVE' directory}
\usage{
get_projects()
}
\value{
characters of project names
}
\description{
Get all possible projects in 'RAVE' directory
}
