% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{global_preferences}
\alias{global_preferences}
\title{Global preferences for pipelines and modules}
\usage{
global_preferences(
  name,
  ...,
  .initial_prefs = list(),
  .overwrite = FALSE,
  .verbose = FALSE
)
}
\arguments{
\item{name}{preference name, must contain only letters, digits,
underscore, and hyphen, will be coerced to lower case (case-insensitive)}

\item{..., .initial_prefs}{key-value pairs of initial preference values}

\item{.overwrite}{whether to overwrite the initial preference values
if they exist.}

\item{.verbose}{whether to verbose the preferences to be saved; default
is false; turn on for debug use}
}
\value{
A persistent map, see \code{\link[dipsaus]{rds_map}}
}
\description{
load persistent global preference settings that can be
accessed across modules, pipelines, and R sessions.
}
\details{
The preferences should not affect how pipeline is working, hence usually
stores minor variables such as graphic options. Changing preferences
will not invalidate pipeline cache.

Developers should maintain and check the preferences at their own risks.
For example, the preferences may not be available. In case of broken files,
please use try-catch clause when trying to access the items.

To avoid performance hit, please do not save functions, environments, only
save atomic items within \code{1 MB}. Though not implemented at this moment,
this restriction will be rigidly enforced in the future.
}
\examples{

if( interactive() ) {
  preference <- global_preferences(
    name = "my_list",
    item1 = "A"
  )

  # get items wrapped in tryCatch
  get_my_preference <- function(name, default = NULL) {
    tryCatch({
      preference$get(name, missing_default = default)
    }, error = function(e) {
      default
    })
  }

  get_my_preference("item1", "missing")
  get_my_preference("item2", "missing")

}

}
