% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{h5_names}
\alias{h5_names}
\title{Returns all names contained in 'HDF5' file}
\usage{
h5_names(file)
}
\arguments{
\item{file, }{'HDF5' file path}
}
\value{
characters, data set names
}
\description{
Returns all names contained in 'HDF5' file
}
