% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rave-prepare-addon-ingest_regressor.R
\name{ingest_regressor}
\alias{ingest_regressor}
\title{Ingest signals according to 'RAVE' repository epoch}
\usage{
ingest_regressor(
  repository,
  source,
  source_type = c("channel", "file", "r_object"),
  filter = NULL,
  ...
)
}
\arguments{
\item{repository}{'RAVE' repository, for example, generated from
\code{\link{prepare_subject_power}}. The repository must contain epoch
information.}

\item{source}{data source, see \code{source_type}}

\item{source_type}{type of the data source, choices are
\describe{
\item{\code{'channel'}}{\code{source} should be interpreted as subject
channels (integer or series of integers)}
\item{\code{'file'}}{\code{source} should be interpreted as file path; the
file format must be either 'Matlab' or 'HDF5', with data names to be the
blocks in the "epoch" file. The sample rate of underlying signals
must coincide with the sample rate presented in the repository}
\item{\code{'r_object'}}{\code{source} should be a list of R objects, where
the names are blocks in the "epoch" file and the data are signal traces;
The sample rate of underlying signals
must coincide with the sample rate presented in the repository}
}}

\item{filter}{\code{NULL} or function; only used if \code{source_type} is
\code{'channel'}. For function,
\code{filter} must have two exact arguments, with the first argument taking
the signal data matrix (time point by channel) and second argument taking
sample rates of length 2. The first sample rate is the original sampling
frequency of the data; the second sample rate is sampling frequency
derived from the repository.}

\item{...}{passed to internal functions.}
}
\value{
A matrix of time by trial
}
\description{
Ingest signals according to 'RAVE' repository epoch
}
