% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{install_modules}
\alias{install_modules}
\title{Install 'RAVE' modules}
\usage{
install_modules(modules, dependencies = FALSE)
}
\arguments{
\item{modules}{a vector of characters, repository names; default is to
automatically determined from a public registry}

\item{dependencies}{whether to update dependent packages; default is false}
}
\value{
nothing
}
\description{
Install 'RAVE' modules
}
