% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is.blank}
\alias{is.blank}
\title{Check If Input Has Blank String}
\usage{
is.blank(x)
}
\arguments{
\item{x}{input data: a vector or an array}
}
\value{
\code{x == ""}
}
\description{
Check If Input Has Blank String
}
