% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is.zerolenth}
\alias{is.zerolenth}
\title{Check If Input Has Zero Length}
\usage{
is.zerolenth(x)
}
\arguments{
\item{x}{input data: a vector, list, or array}
}
\value{
whether \code{x} has zero length
}
\description{
Check If Input Has Zero Length
}
