% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-electrodes-continuous.R
\name{LFP_electrode}
\alias{LFP_electrode}
\title{Definition for electrodes that collects 'LFP' signals}
\description{
Definition for electrodes that collects 'LFP' signals

Definition for electrodes that collects 'LFP' signals
}
\section{Super classes}{
\code{\link[raveio:RAVEAbstarctElectrode]{raveio::RAVEAbstarctElectrode}} -> \code{\link[raveio:Continuous_Electrode]{raveio::Continuous_Electrode}} -> \code{LFP_electrode}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{type of signals collected by the electrode}
}
\if{html}{\out{</div>}}
}
