% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rave-meta.R
\name{load_meta2}
\alias{load_meta2}
\title{Load 'RAVE' subject meta data}
\usage{
load_meta2(meta_type, project_name, subject_code, subject_id, meta_name)
}
\arguments{
\item{meta_type}{electrodes, epochs, time_points, frequencies, references ...}

\item{project_name}{project name}

\item{subject_code}{subject code}

\item{subject_id}{"project_name/subject_code"}

\item{meta_name}{only used if meta_type is epochs or references}
}
\value{
A data frame of the specified meta type or \code{NULL} is no meta
data is found.
}
\description{
Load 'RAVE' subject meta data
}
