% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{load_yaml}
\alias{load_yaml}
\title{A port to \code{\link[yaml]{read_yaml}}}
\usage{
load_yaml(file, ..., map = NULL)
}
\arguments{
\item{file, ...}{passed to \code{\link[yaml]{read_yaml}}}

\item{map}{\code{\link[dipsaus]{fastmap2}} instance or \code{NULL}}
}
\value{
A \code{\link[dipsaus]{fastmap2}}. If \code{map} is provided
then return map, otherwise return newly created one
}
\description{
A port to \code{\link[yaml]{read_yaml}}
}
