% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-variable-collection.R
\name{new_variable_collection}
\alias{new_variable_collection}
\title{Create a collection of constraint variables}
\usage{
new_variable_collection(name = "", explicit = TRUE)
}
\arguments{
\item{name}{collection name, default is empty}

\item{explicit}{whether setting and getting variables should be explicit,
default is \code{TRUE}, which means trying to get undefined variables
will result in errors}
}
\value{
A \code{RAVEVariableCollection} instance
}
\description{
Create a collection of constraint variables
}
\examples{

collection <- new_variable_collection()

# Add unconstrained variables
collection$add_variable(id = "title", "Voltage traces")

# Add a variable with placeholder
collection$add_variable(id = "baseline_window")

# Add variable with constraints
collection$add_variable(
  id = "analysis_range",
  var = new_constrained_variable(
    name = "Analysis range",
    initial_value = c(0, 1),
    constraints = "numeric",
    any.missing = FALSE,
    len = 2,
    sorted = TRUE,
    null.ok = FALSE
  )
)

collection$use_constraints(function(x) {
  missing_values <- vapply(x, inherits, FALSE, "key_missing")
  missing_keys <- names(x)[ missing_values ]
  if(!length(missing_keys)) { return(TRUE) }
  sprintf("Variable [\%s] missing", paste(missing_keys, collapse = ","))
})


# validation will fail
collection$validate(on_error = "message")

# Fix the issue
collection$set_variable("baseline_window", c(-1,0))
collection$validate()

# Get variable values
collection$as_list()
collection[]

# get one variable
collection[["baseline_window"]]

# get partial variables
collection["baseline_window", "analysis_range"]
collection[c("baseline_window", "analysis_range")]

\dontrun{

# error out when explicit
collection[["unregistered_variable"]]
collection[["unregistered_variable"]] <- 1

}

# turn off explicit variable option
collection$explicit <- FALSE
collection[["unregistered_variable"]]
collection[["unregistered_variable"]] <- 1

}
