% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-pipeline_collection.R
\name{pipeline_collection}
\alias{pipeline_collection}
\title{Combine and execute pipelines}
\usage{
pipeline_collection(root_path = NULL, overwrite = FALSE)
}
\arguments{
\item{root_path}{directory to store pipelines and results}

\item{overwrite}{whether to overwrite if \code{root_path} exists; default is
false, and raises an error when \code{root_path} exists}
}
\value{
A \code{\link{PipelineCollections}} instance
}
\description{
Combine and execute pipelines
}
