% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rave-prepare.R
\name{prepare_power}
\alias{prepare_power}
\title{Prepare 'RAVE' subject power (over frequency and time) data}
\usage{
prepare_power(subject, electrodes, epoch_name, reference_name, time_windows)
}
\arguments{
\item{subject}{character of project and subject, such as \code{"demo/YAB"},
or \code{\link{RAVESubject}} instance}

\item{electrodes}{integer vector of electrodes}

\item{epoch_name}{epoch name to be loaded}

\item{reference_name}{reference settings to be loaded}

\item{time_windows}{a list of time windows that are relative to epoch onset
time}
}
\value{
A \code{\link[dipsaus]{fastmap2}} (basically a list) of objects:
\describe{
\item{\code{subject}}{A \code{\link{RAVESubject}} instance}
\item{\code{epoch_name}}{Same as input \code{epoch_name}}
\item{\code{epoch}}{A \code{\link{RAVEEpoch}} instance}
\item{\code{reference_name}}{Same as input \code{reference_name}}
\item{\code{reference_table}}{A data frame of reference}
\item{\code{electrode_table}}{A data frame of electrode information}
\item{\code{frequency}}{A vector of frequencies}
\item{\code{time_points}}{A vector of time-points}
\item{\code{power_list}}{A list of power data of the electrodes}
\item{\code{power_dimnames}}{A list of trial indices, frequencies, time
points, and electrodes that are loaded}
}
}
\description{
Prepare 'RAVE' subject power (over frequency and time) data
}
