% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress_with_logger}
\alias{progress_with_logger}
\title{Enhanced progress with logger message}
\usage{
progress_with_logger(
  title,
  max = 1,
  ...,
  quiet = FALSE,
  session = shiny::getDefaultReactiveDomain(),
  shiny_auto_close = FALSE,
  outputId = NULL,
  log
)
}
\arguments{
\item{title, max, ..., quiet, session, shiny_auto_close}{see
\code{\link[dipsaus]{progress2}}}

\item{outputId}{will be used if package \code{'shidashi'} is installed,
otherwise will be ignored}

\item{log}{function, \code{NULL}, or missing; default is missing, which
will use \code{logger} function in the package \code{'ravedash'}, or
\code{\link[dipsaus]{cat2}} if \code{'ravedash'} is not installed. If
\code{log=NULL}, then the message will be suppressed in 'shiny' applications.
If a function provided, then the function will be called.}
}
\value{
A list, see \code{\link[dipsaus]{progress2}}
}
\description{
For best performance, please install \code{'ravedash'}. This
function can replace \code{\link[dipsaus]{progress2}}.
}
