% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rave-import.R
\name{rave_directories}
\alias{rave_directories}
\title{Returns a list of 'RAVE' directories}
\usage{
rave_directories(
  subject_code,
  project_name,
  blocks = NULL,
  .force_format = c("", "native", "BIDS")
)
}
\arguments{
\item{subject_code}{'RAVE' subject code}

\item{project_name}{'RAVE' project name}

\item{blocks}{session or block names, optional}

\item{.force_format}{format of the data, default is automatically detected.}
}
\value{
A list of directories
}
\description{
This function is internally used and should not be called
directly.
}
