% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rave-import.R
\name{rave_import}
\alias{rave_import}
\title{Import data into 'rave'}
\usage{
rave_import(
  project_name,
  subject_code,
  blocks,
  electrodes,
  format,
  sample_rate,
  data_type = "lfp",
  ...
)
}
\arguments{
\item{project_name}{project name, for 'rave' native structure, this can be
any character; for 'BIDS' format, this must be consistent with 'BIDS'
project name. For subjects with multiple tasks, see Section "'RAVE' Project"}

\item{subject_code}{subject code in character. For 'rave' native structure,
this is a folder name under raw directory. For 'BIDS', this is subject
label without \code{"sub-"} prefix}

\item{blocks}{characters, for 'rave' native format, this is the folder names
subject directory; for 'BIDS', this is session name with \code{"ses-"}.
Section "Block vs. Session" for different meaning of "blocks" in 'rave'
and 'BIDS'}

\item{electrodes}{integers electrode numbers}

\item{format}{integer from 1 to 6, or character. For characters, you can get
options by running \code{names(LFP_FORMATS)}}

\item{sample_rate}{sample frequency, must be positive}

\item{data_type}{electrode type; only \code{'lfp'} is supported}

\item{...}{other parameters, depend on format applied. For 'BIDS' formats,
please specify \code{task_runs} variable; see Section "Block vs. Session"
and Section "File Formats"}
}
\description{
Import files with pre-defined structures. Supported file
formats include 'Matlab', 'HDF5', 'EDF(+)', 'BrainVision'
(\code{'.eeg/.dat/.vhdr'}). Supported file structures include 'rave' native
structure and 'BIDS' format.
}
\section{'RAVE' Project}{

A 'rave' project can be very flexible. A project can refer to a task, a
research objective, or "arbitrarily" as long as you find common research
interests among subjects. One subject can appear in multiple projects with
different blocks, hence \code{project_name} should be task-based or
objective-based. There is no concept of "project" in 'rave' raw directory.
When importing data, you choose subset of blocks from subjects forming
a project.
This is different from 'BIDS' format. In 'BIDS',
one subject can be only a sub-folder of project, and a project must be
present in raw format.

This two implementations have their own pros and cons. However, when
importing 'BIDS' data into 'rave', \code{project_name} must be consistent
with 'BIDS' project name as a compromise. Once imported, you may change
the project folder name in imported rave data directory to other names.
Because once raw traces are imported, 'rave' data will become self-contained
and 'BIDS' data are no longer required for analysis.
This naming inconsistency will also be ignored.
}

\section{Block vs. Session}{

\code{blocks} has different meaning in 'rave' and 'BIDS'. In 'rave', it means
combination of session (days), task, and run, i.e. a block of continuous
signals captured. Raw data files are supposed to be stored in file
hierarchy of \code{<raw-root>/<subject_code>/<block>/<datafiles>}.
However, this structure is incompatible with 'BIDS'.
In 'BIDS' sessions, tasks, and runs are separated, and only session names
are indicated under subject folder.
Therefore when importing data from 'BIDS' format, \code{block} argument
needs to be session names to comply with \code{'sub/session'} structure.
To record "real" blocks, \code{task_runs} must be specified.

For 'rave' native raw data format, \code{block} will be as-is once imported.
\cr
For 'BIDS' format, \code{task_runs} will be treated as blocks once imported.
}

\section{File Formats}{

Depending on format supported, \code{...} has the following arguments:
\describe{
\item{\code{format=1}, or \code{".mat/.h5 file per electrode per block"}}{\code{...} are ignored}
\item{\code{format=2}, or \code{"Single .mat/.h5 file per block"}}{\code{...} are ignored}
\item{\code{format=3}, or \code{"Single EDF(+) file per block"}}{
\code{convertion} (optional), convert electric potential unit;
choices are \code{'NA'} (no conversion), \code{'V'} (volt), \code{'mV'}
(milli-volt), and \code{'uV'} (micro-volt)
}
\item{\code{format=4}, or \code{"Single BrainVision file (.vhdr+.eeg, .vhdr+.dat) per block"}}{
\code{...} are ignored
}
\item{\code{format=5}, or \code{"BIDS & EDF(+)"}}{
\code{task_runs} (mandatory), characters, combination of session, task name,
and run number. For example, a task header file in BIDS with name
\code{'sub-som682_ses-01_task-visual_run-01_ieeg.json'} has \code{task_runs}
name as \code{'01-visual-01'}, where the first \code{'01'} refers to session,
\code{'visual'} is task name, and the second \code{'01'} is run number.
\cr
\code{convertion} (optional), convert electric potential unit; see
previous format: \code{"Single EDF(+) file per block"}
}
\item{\code{format=6}, or \code{"BIDS & BrainVision (.vhdr+.eeg, .vhdr+.dat)"}}{
\code{task_runs} (mandatory), characters, see previous format:
\code{"BIDS & EDF(+)"}
}
}
}

