% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward-compatibility.R
\name{rave_subject_format_conversion}
\alias{rave_subject_format_conversion}
\title{Compatibility support for 'RAVE' 1.0 format}
\usage{
rave_subject_format_conversion(subject, verbose = TRUE, ...)
}
\arguments{
\item{subject}{'RAVE' subject characters, such as \code{'demo/YAB'}, or
a subject instance generated from \code{\link{RAVESubject}}}

\item{verbose}{whether to verbose the messages}

\item{...}{ignored, reserved for future use}
}
\value{
Nothing
}
\description{
Convert 'RAVE' subject generated by 2.0 pipeline such that
1.0 modules can use the data. The subject must have valid electrodes.
The data must be imported, with time-frequency transformed to pass the
validation before converting.
}
