% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-electrodes.R
\name{RAVEAbstarctElectrode}
\alias{RAVEAbstarctElectrode}
\title{Abstract definition of electrode class in RAVE}
\description{
Abstract definition of electrode class in RAVE

Abstract definition of electrode class in RAVE
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{type of electrode}

\item{\code{subject}}{subject instance (\code{\link{RAVESubject}})}

\item{\code{number}}{integer stands for electrode number or reference ID}

\item{\code{reference}}{reference electrode, either \code{NULL} for no reference
or an electrode instance inherits \code{RAVEAbstarctElectrode}}

\item{\code{epoch}}{a \code{\link{RAVEEpoch}} instance}

\item{\code{cached_reference}}{character, refer to current cached reference}

\item{\code{is_reference}}{whether this instance is a reference electrode}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{exists}}{whether electrode exists in subject}

\item{\code{preprocess_file}}{path to preprocess 'HDF5' file}

\item{\code{power_file}}{path to power 'HDF5' file}

\item{\code{power_cached}}{path to power 'FST' file}

\item{\code{phase_file}}{path to phase 'HDF5' file}

\item{\code{voltage_file}}{path to voltage 'HDF5' file}

\item{\code{reference_name}}{reference electrode name (field: number)}

\item{\code{cache_path}}{run-time cache path; \code{NA} if epoch is missing}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-.set_cachename}{\code{RAVEAbstarctElectrode$.set_cachename()}}
\item \href{#method-new}{\code{RAVEAbstarctElectrode$new()}}
\item \href{#method-.set_reference}{\code{RAVEAbstarctElectrode$.set_reference()}}
\item \href{#method-set_epoch}{\code{RAVEAbstarctElectrode$set_epoch()}}
\item \href{#method-clear_cache}{\code{RAVEAbstarctElectrode$clear_cache()}}
\item \href{#method-clear_memory}{\code{RAVEAbstarctElectrode$clear_memory()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.set_cachename"></a>}}
\if{latex}{\out{\hypertarget{method-.set_cachename}{}}}
\subsection{Method \code{.set_cachename()}}{
set cache name, internally used
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVEAbstarctElectrode$.set_cachename(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character, internally used}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVEAbstarctElectrode$new(subject, number, is_reference = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subject}}{character or \code{\link{RAVESubject}} instance}

\item{\code{number}}{current electrode number or reference ID}

\item{\code{is_reference}}{whether instance is a reference}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.set_reference"></a>}}
\if{latex}{\out{\hypertarget{method-.set_reference}{}}}
\subsection{Method \code{.set_reference()}}{
set reference for instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVEAbstarctElectrode$.set_reference(reference, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reference}}{\code{NULL} or \code{RAVEAbstarctElectrode} instance}

\item{\code{type}}{reference electrode type, default is the same as current
instance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_epoch"></a>}}
\if{latex}{\out{\hypertarget{method-set_epoch}{}}}
\subsection{Method \code{set_epoch()}}{
set epoch instance for the electrode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVEAbstarctElectrode$set_epoch(epoch)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{epoch}}{characters or \code{\link{RAVEEpoch}} instance. For
characters, make sure \code{"epoch_<name>.csv"} is in meta folder.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_cache"></a>}}
\if{latex}{\out{\hypertarget{method-clear_cache}{}}}
\subsection{Method \code{clear_cache()}}{
method to clear cache on hard drive
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVEAbstarctElectrode$clear_cache(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to child instances}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_memory"></a>}}
\if{latex}{\out{\hypertarget{method-clear_memory}{}}}
\subsection{Method \code{clear_memory()}}{
method to clear memory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAVEAbstarctElectrode$clear_memory(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to child instances}
}
\if{html}{\out{</div>}}
}
}
}
