% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{raveio-option}
\alias{raveio-option}
\alias{raveio_setopt}
\alias{raveio_resetopt}
\alias{raveio_getopt}
\title{Set/Get 'raveio' option}
\usage{
raveio_setopt(key, value)

raveio_resetopt(all = FALSE)

raveio_getopt(key, default = NA)
}
\arguments{
\item{key}{character, option name}

\item{value}{character or logical of length 1, option value}

\item{all}{whether to reset all non-default keys}

\item{default}{is key not found, return default value}
}
\description{
Persist settings on local configuration file
}
\details{
\code{raveio_setopt} stores key-value pair in local path.
The values are persistent and shared across multiple sessions.
There are some read-only keys such as \code{"session_string"}. Trying to
set those keys will result in error.

\code{raveio_getopt} returns value corresponding to the keys. If key is
missing, the whole option will be returned.

If set \code{all=TRUE}, \code{raveio_resetopt} resets all keys including
non-standard ones. However \code{"session_string"} will never reset.
}
\seealso{
\code{R_user_dir}
}
