% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{read-write-fst}
\alias{read-write-fst}
\alias{save_fst}
\alias{load_fst}
\title{Read a 'fst' file}
\usage{
save_fst(x, path, ...)

load_fst(path, ..., as.data.table = TRUE)
}
\arguments{
\item{x}{data frame to write to path}

\item{path}{path to 'fst' file: must not be connection.}

\item{...}{passed to \code{read_fst} or \code{\link[fst]{write_fst}}}

\item{as.data.table}{passed to \code{read_fst} in \code{fst} package}
}
\description{
Read a 'fst' file
}
