% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{read_csv_no_header}
\alias{read_csv_no_header}
\title{Read csv file and ignore headers}
\usage{
read_csv_no_header(file, nrows = Inf, drop = NULL)
}
\arguments{
\item{file}{csv file to read from. The file must contains all
numerical values}

\item{nrows}{number of rows to read}

\item{drop}{passed to \code{\link[data.table]{fread}}}
}
\description{
Resolved some iEEG format import situation when
the header could be missing.
}
\details{
The function checks the first two rows of csv file.
If the first row has different \code{\link{storage.mode}} than
the second row, then the first row is considered header, otherwise
header is treated missing. Note \code{file} must have at least two
rows.
}
