% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edf.R
\name{read_edf_signal}
\alias{read_edf_signal}
\title{Read 'EDF(+)' or 'BDF(+)' file signals}
\usage{
read_edf_signal(
  path,
  signal_numbers = NULL,
  convert_volt = c("NA", "V", "mV", "uV")
)
}
\arguments{
\item{path}{file path, passed to \code{readEdfHeader}}

\item{signal_numbers}{channel/electrode numbers}

\item{convert_volt}{convert voltage (electric potential) to a new unit,
\code{NA} means no conversion, other choices are \code{'V'}, \code{'mV'}, and
\code{'uV'}.}
}
\value{
A list containing header information, signal lists, and
channel/electrode names. If \code{signal_numbers} is specified,
the corresponding names should appear as \code{selected_signal_names}.
\code{get_signal()} can get physical signals after unit conversion.
}
\description{
Read 'EDF(+)' or 'BDF(+)' file signals
}
