% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-ieegio.R, R/reexport-ravecore.R,
%   R/reexport-ravepipeline.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{LazyH5}
\alias{LazyFST}
\alias{convert_fst_to_hdf5}
\alias{convert_fst_to_csv}
\alias{RAVEAbstarctElectrode}
\alias{new_electrode}
\alias{new_reference}
\alias{RAVEEpoch}
\alias{LFP_electrode}
\alias{Spike_electrode}
\alias{Auxiliary_electrode}
\alias{LFP_reference}
\alias{RAVEPreprocessSettings}
\alias{cmd_execute}
\alias{as_rave_project}
\alias{get_projects}
\alias{RAVEProject}
\alias{as_rave_subject}
\alias{RAVESubject}
\alias{YAELProcess}
\alias{as_yael_process}
\alias{prepare_subject_bare0}
\alias{prepare_subject_bare}
\alias{prepare_subject_with_epochs}
\alias{prepare_subject_raw_voltage_with_epochs}
\alias{prepare_subject_voltage_with_epochs}
\alias{prepare_subject_time_frequency_coefficients_with_epochs}
\alias{prepare_subject_power}
\alias{prepare_subject_power_with_epochs}
\alias{prepare_subject_phase_with_epochs}
\alias{prepare_subject_with_blocks}
\alias{prepare_subject_voltage_with_blocks}
\alias{prepare_subject_time_frequency_coefficients_with_blocks}
\alias{prepare_subject_power_with_blocks}
\alias{prepare_subject_phase_with_blocks}
\alias{archive_subject}
\alias{install_subject}
\alias{generate_reference}
\alias{cache_root}
\alias{backup_file}
\alias{clear_cached_files}
\alias{rave_brain}
\alias{transform_point_to_template}
\alias{transform_thinfilm_to_mni152}
\alias{compose_channel}
\alias{validate_time_window}
\alias{load_meta2}
\alias{save_meta2}
\alias{validate_subject}
\alias{cmd_run_r}
\alias{SIGNAL_TYPES}
\alias{LOCATION_TYPES}
\alias{MNI305_to_MNI152}
\alias{YAEL_IMAGE_TYPES}
\alias{IMPORT_FORMATS}
\alias{export_table}
\alias{import_table}
\alias{normalize_commandline_path}
\alias{cmd_dcm2niix}
\alias{cmd_freesurfer_home}
\alias{cmd_fsl_home}
\alias{cmd_afni_home}
\alias{cmd_homebrew}
\alias{rscript_path}
\alias{yael_preprocess}
\alias{cmd_run_yael_preprocess}
\alias{cmd_run_freesurfer_recon_all}
\alias{cmd_run_freesurfer_recon_all_clinical}
\alias{ants_coreg}
\alias{cmd_run_ants_coreg}
\alias{cmd_run_dcm2niix}
\alias{niftyreg_coreg}
\alias{cmd_run_niftyreg_coreg}
\alias{dir_create2}
\alias{load_yaml}
\alias{save_yaml}
\alias{glue}
\alias{raveio_setopt}
\alias{raveio_getopt}
\alias{raveio_resetopt}
\alias{raveio_confpath}
\alias{PipelineTools}
\alias{PipelineResult}
\alias{pipeline}
\alias{pipeline_from_path}
\alias{pipeline_install_local}
\alias{pipeline_install_github}
\alias{pipeline_root}
\alias{pipeline_list}
\alias{pipeline_find}
\alias{pipeline_attach}
\alias{pipeline_run}
\alias{pipeline_clean}
\alias{pipeline_run_bare}
\alias{pipeline_eval}
\alias{pipeline_render}
\alias{configure_knitr}
\alias{pipeline_setup_rmd}
\alias{load_targets}
\alias{pipeline_target_names}
\alias{pipeline_debug}
\alias{pipeline_dep_targets}
\alias{pipeline_visualize}
\alias{pipeline_progress}
\alias{pipeline_fork}
\alias{pipeline_build}
\alias{pipeline_read}
\alias{pipeline_vartable}
\alias{pipeline_hasname}
\alias{pipeline_watch}
\alias{pipeline_create_template}
\alias{pipeline_create_subject_pipeline}
\alias{pipeline_description}
\alias{pipeline_settings_set}
\alias{pipeline_settings_get}
\alias{pipeline_load_extdata}
\alias{pipeline_save_extdata}
\alias{pipeline_shared}
\alias{pipeline_set_preferences}
\alias{pipeline_get_preferences}
\alias{pipeline_has_preferences}
\alias{pipeline_collection}
\alias{PipelineCollections}
\alias{load_snippet}
\alias{update_local_snippet}
\alias{install_snippet}
\alias{list_snippets}
\alias{module_add}
\alias{install_modules}
\alias{module_registry}
\alias{module_registry2}
\alias{get_modules_registries}
\alias{get_module_description}
\alias{add_module_registry}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ieegio}{\code{\link[ieegio:convert-fst]{convert_fst_to_csv}}, \code{\link[ieegio:convert-fst]{convert_fst_to_hdf5}}, \code{\link[ieegio]{LazyFST}}, \code{\link[ieegio]{LazyH5}}}

  \item{ravecore}{\code{\link[ravecore:cmd_run_ants_coreg]{ants_coreg}}, \code{\link[ravecore]{archive_subject}}, \code{\link[ravecore]{as_rave_project}}, \code{\link[ravecore:new_rave_subject]{as_rave_subject}}, \code{\link[ravecore]{as_yael_process}}, \code{\link[ravecore]{Auxiliary_electrode}}, \code{\link[ravecore]{backup_file}}, \code{\link[ravecore:cache_path]{cache_root}}, \code{\link[ravecore:cache_path]{clear_cached_files}}, \code{\link[ravecore:rave_cmd_tools]{cmd_afni_home}}, \code{\link[ravecore:rave_cmd_tools]{cmd_dcm2niix}}, \code{\link[ravecore:cmd-external]{cmd_execute}}, \code{\link[ravecore:rave_cmd_tools]{cmd_freesurfer_home}}, \code{\link[ravecore:rave_cmd_tools]{cmd_fsl_home}}, \code{\link[ravecore:rave_cmd_tools]{cmd_homebrew}}, \code{\link[ravecore]{cmd_run_ants_coreg}}, \code{\link[ravecore]{cmd_run_dcm2niix}}, \code{\link[ravecore]{cmd_run_freesurfer_recon_all}}, \code{\link[ravecore:cmd_run_freesurfer_recon_all]{cmd_run_freesurfer_recon_all_clinical}}, \code{\link[ravecore:niftyreg_coreg]{cmd_run_niftyreg_coreg}}, \code{\link[ravecore:cmd-external]{cmd_run_r}}, \code{\link[ravecore]{cmd_run_yael_preprocess}}, \code{\link[ravecore]{compose_channel}}, \code{\link[ravecore]{export_table}}, \code{\link[ravecore]{generate_reference}}, \code{\link[ravecore]{get_projects}}, \code{\link[ravecore:ravecore-constants]{IMPORT_FORMATS}}, \code{\link[ravecore:export_table]{import_table}}, \code{\link[ravecore]{install_subject}}, \code{\link[ravecore]{LFP_electrode}}, \code{\link[ravecore]{LFP_reference}}, \code{\link[ravecore:meta-data]{load_meta2}}, \code{\link[ravecore:ravecore-constants]{LOCATION_TYPES}}, \code{\link[ravecore:ravecore-constants]{MNI305_to_MNI152}}, \code{\link[ravecore]{new_electrode}}, \code{\link[ravecore:new_electrode]{new_reference}}, \code{\link[ravecore]{niftyreg_coreg}}, \code{\link[ravecore:rave_cmd_tools]{normalize_commandline_path}}, \code{\link[ravecore]{prepare_subject_bare}}, \code{\link[ravecore:prepare_subject_bare]{prepare_subject_bare0}}, \code{\link[ravecore:prepare_subject_with_blocks]{prepare_subject_phase_with_blocks}}, \code{\link[ravecore:prepare_subject_with_epochs]{prepare_subject_phase_with_epochs}}, \code{\link[ravecore:prepare_subject_with_epochs]{prepare_subject_power}}, \code{\link[ravecore:prepare_subject_with_blocks]{prepare_subject_power_with_blocks}}, \code{\link[ravecore:prepare_subject_with_epochs]{prepare_subject_power_with_epochs}}, \code{\link[ravecore:prepare_subject_with_epochs]{prepare_subject_raw_voltage_with_epochs}}, \code{\link[ravecore:prepare_subject_with_blocks]{prepare_subject_time_frequency_coefficients_with_blocks}}, \code{\link[ravecore:prepare_subject_with_epochs]{prepare_subject_time_frequency_coefficients_with_epochs}}, \code{\link[ravecore:prepare_subject_with_blocks]{prepare_subject_voltage_with_blocks}}, \code{\link[ravecore:prepare_subject_with_epochs]{prepare_subject_voltage_with_epochs}}, \code{\link[ravecore]{prepare_subject_with_blocks}}, \code{\link[ravecore]{prepare_subject_with_epochs}}, \code{\link[ravecore]{rave_brain}}, \code{\link[ravecore]{RAVEAbstarctElectrode}}, \code{\link[ravecore]{RAVEEpoch}}, \code{\link[ravecore]{RAVEPreprocessSettings}}, \code{\link[ravecore]{RAVEProject}}, \code{\link[ravecore]{RAVESubject}}, \code{\link[ravecore:rave_cmd_tools]{rscript_path}}, \code{\link[ravecore:meta-data]{save_meta2}}, \code{\link[ravecore:ravecore-constants]{SIGNAL_TYPES}}, \code{\link[ravecore]{Spike_electrode}}, \code{\link[ravecore]{transform_point_to_template}}, \code{\link[ravecore:transform_point_to_template]{transform_thinfilm_to_mni152}}, \code{\link[ravecore]{validate_subject}}, \code{\link[ravecore]{validate_time_window}}, \code{\link[ravecore:ravecore-constants]{YAEL_IMAGE_TYPES}}, \code{\link[ravecore:cmd_run_yael_preprocess]{yael_preprocess}}, \code{\link[ravecore]{YAELProcess}}}

  \item{ravepipeline}{\code{\link[ravepipeline:module_registry]{add_module_registry}}, \code{\link[ravepipeline:pipeline-knitr-markdown]{configure_knitr}}, \code{\link[ravepipeline]{dir_create2}}, \code{\link[ravepipeline:module_registry]{get_module_description}}, \code{\link[ravepipeline:module_registry]{get_modules_registries}}, \code{\link[ravepipeline:reexports]{glue}}, \code{\link[ravepipeline]{install_modules}}, \code{\link[ravepipeline:rave-snippet]{install_snippet}}, \code{\link[ravepipeline:rave-snippet]{list_snippets}}, \code{\link[ravepipeline:rave-snippet]{load_snippet}}, \code{\link[ravepipeline:rave-pipeline]{load_targets}}, \code{\link[ravepipeline:read-write-yaml]{load_yaml}}, \code{\link[ravepipeline]{module_add}}, \code{\link[ravepipeline]{module_registry}}, \code{\link[ravepipeline:module_registry]{module_registry2}}, \code{\link[ravepipeline]{pipeline}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_attach}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_build}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_clean}}, \code{\link[ravepipeline]{pipeline_collection}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_create_subject_pipeline}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_create_template}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_debug}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_dep_targets}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_description}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_eval}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_find}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_fork}}, \code{\link[ravepipeline:pipeline]{pipeline_from_path}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_get_preferences}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_has_preferences}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_hasname}}, \code{\link[ravepipeline:pipeline_install]{pipeline_install_github}}, \code{\link[ravepipeline:pipeline_install]{pipeline_install_local}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_list}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_load_extdata}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_progress}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_read}}, \code{\link[ravepipeline:pipeline-knitr-markdown]{pipeline_render}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_root}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_run}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_run_bare}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_save_extdata}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_set_preferences}}, \code{\link[ravepipeline:pipeline_settings_get_set]{pipeline_settings_get}}, \code{\link[ravepipeline:pipeline_settings_get_set]{pipeline_settings_set}}, \code{\link[ravepipeline:pipeline-knitr-markdown]{pipeline_setup_rmd}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_shared}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_target_names}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_vartable}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_visualize}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_watch}}, \code{\link[ravepipeline]{PipelineCollections}}, \code{\link[ravepipeline]{PipelineResult}}, \code{\link[ravepipeline]{PipelineTools}}, \code{\link[ravepipeline:raveio-option]{raveio_confpath}}, \code{\link[ravepipeline:raveio-option]{raveio_getopt}}, \code{\link[ravepipeline:raveio-option]{raveio_resetopt}}, \code{\link[ravepipeline:raveio-option]{raveio_setopt}}, \code{\link[ravepipeline:read-write-yaml]{save_yaml}}, \code{\link[ravepipeline:rave-snippet]{update_local_snippet}}}
}}

