% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{rstudio_active_project}
\alias{rstudio_active_project}
\title{Get 'RStudio' active project}
\usage{
rstudio_active_project()
}
\value{
If 'RStudio' is running and current project is not none, return
project name, otherwise return \code{NULL}
}
\description{
Get 'RStudio' active project
}
