% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{rstudio_viewer}
\alias{rstudio_viewer}
\title{Get 'RStudio' Viewer, or Return Default}
\usage{
rstudio_viewer(..., default = TRUE)
}
\arguments{
\item{...}{passed to \code{\link[rstudioapi]{viewer}}}

\item{default}{if \code{\link{verify_rstudio_version}} fails, the
value to return. Default is \code{TRUE}}
}
\value{
If \code{\link[rstudioapi]{viewer}} can be called and
'RStudio' is running, then launch 'RStudio' internal viewer.
Otherwise if \code{default} is a function such as
\code{\link[utils]{browseURL}}, then call the function with given
arguments. If \code{default} is not a function, return \code{default}
}
\description{
Get 'RStudio' Viewer, or Return Default
}
