% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{safe_write_csv}
\alias{safe_write_csv}
\title{Save data to comma separated value files with backups}
\usage{
safe_write_csv(x, file, ..., quiet = FALSE)
}
\arguments{
\item{x, file, ...}{pass to \code{\link[utils]{write.csv}}}

\item{quiet}{whether to suppress overwrite message}
}
\value{
Normalized path of \code{file}
}
\description{
Save comma separated value files, if file exists,
backup original file.
}
