% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{select_path}
\alias{select_path}
\title{Use 'RStudio' to Select a Path on the Server}
\usage{
select_path(is_directory = TRUE)
}
\arguments{
\item{is_directory}{whether the path should be a directory}
}
\value{
Raise error if \code{\link{verify_rstudio_version}} fails,
otherwise returns the selected path
}
\description{
Use 'RStudio' to Select a Path on the Server
}
