% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{time_diff2}
\alias{time_diff2}
\title{Calculate time difference in seconds}
\usage{
time_diff2(start, end, units = "secs", label = "")
}
\arguments{
\item{start, end}{start and end of timer}

\item{units}{passed to \code{\link[dipsaus]{time_delta}}}

\item{label}{\code{rave-units} label for display purpose.}
}
\value{
A number inherits \code{rave-units} class.
}
\description{
Calculate time difference in seconds
}
\examples{
start <- Sys.time()
Sys.sleep(0.1)
end <- Sys.time()
dif <- time_diff2(start, end, label = 'Running ')
print(dif, digits = 4)

is.numeric(dif)

dif + 1

}
