% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeBrain-instance.R
\name{transform_point_to_template}
\alias{transform_point_to_template}
\title{Calculate template 'MNI' coordinates for points on native brain}
\usage{
transform_point_to_template(
  subject,
  positions,
  space = c("scannerRAS", "tkrRAS"),
  mapping_method = c("volumetric", "surface"),
  flip_hemisphere = FALSE,
  ...
)
}
\arguments{
\item{subject}{'RAVE' subject}

\item{positions}{optional matrix of 3 columns, either in scanner or surface
space (specified by \code{space}); default is missing and will use the
electrode localization results (\code{electrodes.csv})}

\item{space}{if \code{positions} is given, which native coordinate system
should be used; default is native 'T1' (or \code{'scannerRAS'}); alternative
is 'FreeSurfer' surface coordinate (or \code{'tkrRAS'})}

\item{mapping_method}{whether the mapping is \code{'volumetric'} or
\code{'surface'}; default is the former.}

\item{flip_hemisphere}{whether to flip the hemisphere; default is
\code{FALSE}}

\item{...}{passed to other inner functions. For volume mapping, additional
arguments include \code{method} with choices \code{'auto'} (default),
\code{'affine'}, and \code{'nonlinear'}. For surface mapping, additional
arguments include \code{use_surface} (which surface to project to)
with default \code{'pial'}; \code{template} (which template to use for
calculating 'MNI' coordinates) with a default template 'FreeSurfer'
'MNI' brain \code{'cvs_avg35_inMNI152'}}
}
\description{
Calculate template 'MNI' coordinates for points on native brain
}
\examples{

if(interactive()) {

transform_point_to_template('demo/DemoSubject', mapping_method = "volumetric")

}

}
