% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{url_neurosynth}
\alias{url_neurosynth}
\title{Get 'Neurosynth' website address using 'MNI152' coordinates}
\usage{
url_neurosynth(x, y, z)
}
\arguments{
\item{x, y, z}{numerical values: the right-anterior-superior 'RAS'
coordinates in \code{'MNI152'} space}
}
\value{
'Neurosynth' website address
}
\description{
Get 'Neurosynth' website address using 'MNI152' coordinates
}
