% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{verify_rstudio_version}
\alias{verify_rstudio_version}
\title{Verify 'RStudio' Version}
\usage{
verify_rstudio_version(version_needed = "1.2")
}
\arguments{
\item{version_needed}{minimum version required}
}
\value{
whether 'RStudio' is running and its version is above the
required
}
\description{
Verify 'RStudio' Version
}
