% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-preprocess.R
\name{yael-image-processing}
\alias{yael-image-processing}
\alias{yael_preprocess_t1}
\alias{yael_coregistration}
\alias{yael_preprocess}
\alias{yael_postprocess}
\title{'YAEL' imaging processing}
\usage{
yael_preprocess_t1(
  subject_code,
  t1_path,
  templates = "mni_icbm152_nlin_asym_09a",
  freesurfer = NA,
  verbose = TRUE
)

yael_coregistration(subject_code, ct_path, t1_path = NA, verbose = TRUE)

yael_preprocess(
  subject_code,
  t1_path,
  ct_path,
  templates = "mni_icbm152_nlin_asym_09a",
  freesurfer = NA,
  verbose = TRUE,
  parallel = TRUE,
  launch_rave = FALSE,
  wait_recon = FALSE
)

yael_postprocess(
  subject_code,
  roi_folder = NULL,
  templates = "mni_icbm152_nlin_asym_09a",
  verbose = TRUE
)
}
\arguments{
\item{subject_code}{'RAVE' subject ID or subject code (default project will
be \code{'YAEL'})}

\item{t1_path}{path to 'T1' image}

\item{templates}{template image to align 'T1'; currently supports
\code{'mni_icbm152_nlin_asym_09a'} and \code{'mni_icbm152_nlin_asym_09b'}}

\item{freesurfer}{whether to run 'FreeSurfer' command, default is
automatically determined; will enable on 'Linux' and 'OSX' and when the
image directory is empty, and disable on 'Windows' or the 'FreeSurfer' folder
has been previously generated.}

\item{verbose}{whether to verbose the progress; default is true}

\item{ct_path}{path to 'CT' image}

\item{parallel}{whether to run registration and 'T1' processing at the same
time; default is false. Turn it on to save time, but only available when
running in 'RStudio'}

\item{launch_rave}{whether to launch 'RAVE' after process}

\item{wait_recon}{whether to wait for 'FreeSurfer' reconstruction to finish
before launching 'RAVE'}

\item{roi_folder}{'ROI' or atlas folder in template}
}
\value{
\code{yael_preprocess_t1} and \code{yael_preprocess} returns 'bash'
command if 'FreeSurfer' is not available, or launches 'FreeSurfer' in the
background if available.
}
\description{
For preliminary process, aligns 'T1' to template image, performs 'FreeSurfer'
reconstruction (if supported), register 'MRI' to 'CT'; for post-process,
calculate the electrode coordinates in template coordinate system, and
map atlases or 'ROI' to native brain.
}
\examples{

\dontrun{

library(raveio)
yael_preprocess(
  subject_code = "YAB",
  t1_path = "~/YAB_T1.nii",
  ct_path = "~/YAB_CT.nii",
  templates = "mni_icbm152_nlin_asym_09a",
  launch_rave = TRUE
)

# run after localization
yael_postprocess(
  subject_code = "YAB",
  roi_folder = "~/template_roi/harvardox_amygdala_R_bin.nii.gz",
  templates = "mni_icbm152_nlin_asym_09a"
)


}

}
