% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{catgl}
\alias{catgl}
\title{Print colored messages}
\usage{
catgl(..., .envir = parent.frame(), level = "DEBUG", .pal)
}
\arguments{
\item{..., .envir}{passed to \code{\link[glue]{glue}}}

\item{level}{passed to \code{\link[dipsaus]{cat2}}}

\item{.pal}{see \code{pal} in \code{\link[dipsaus]{cat2}}}
}
\description{
Print colored messages
}
