% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-BlackrockFile.R
\name{convert_blackrock}
\alias{convert_blackrock}
\title{Convert 'BlackRock' 'NEV/NSx' files}
\usage{
convert_blackrock(
  file,
  block = NULL,
  subject = NULL,
  to = NULL,
  comments = TRUE,
  format = c("mat", "hdf5")
)
}
\arguments{
\item{file}{path to any 'NEV/NSx' file}

\item{block}{the block name, default is file name}

\item{subject}{subject code to save the files; default is \code{NULL}}

\item{to}{save to path, must be a directory; default is under the file path.
If \code{subject} is provided, then the default is \code{subject} raw
directory path}

\item{comments}{whether to extract comment section as epoch; default is true}

\item{format}{output format, choices are \code{'mat'} or \code{'hdf5'}}
}
\value{
The results will be stored in directory specified by \code{to}.
Please read the output message carefully.
}
\description{
Convert 'BlackRock' 'NEV/NSx' files
}
