% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{find_path}
\alias{find_path}
\title{Try to find path along the root directory}
\usage{
find_path(path, root_dir)
}
\arguments{
\item{path}{path to a file. It's fine if the file is missing}

\item{root_dir}{root directory of the file}
}
\value{
The absolute path of file if exists, or \code{NULL} if
missing/failed.
}
\description{
Try to find \code{path} under root directory even
if the original path is missing
}
\details{
When file is absent, \code{find_path} concatenates the
root directory and path combined to find the file. For example,
if the root directory is \code{"~/"}, and path is \code{"a/b/c/d"},
the function first seek for existence of \code{"~/a/b/c/d"}. If failed,
then \code{"~/b/c/d"}, and then \code{"~/c/d"} until reaching
top (root directory).
}
\examples{
\dontrun{
# This example runs when demo (YAB) data are installed

# Case 1: path exists from root directory
find_path('demo/YAB/rave/meta/electrodes.csv',
          root_dir = '~/rave_data/data_dir')

# Case 2: path missing from root directory
find_path('random/folder/not/exists/demo/YAB/rave/meta/electrodes.csv',
          root_dir = '~/rave_data/data_dir')

}


}
