% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-project.R
\name{get_projects}
\alias{get_projects}
\title{Get all possible projects in 'RAVE' directory}
\usage{
get_projects(refresh = TRUE)
}
\arguments{
\item{refresh}{whether to refresh the cache; default is true}
}
\value{
characters of project names
}
\description{
Get all possible projects in 'RAVE' directory
}
