% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rave-meta.R
\name{import_electrode_table}
\alias{import_electrode_table}
\title{Import electrode table into subject meta folder}
\usage{
import_electrode_table(path, subject, use_fs = NA, dry_run = FALSE, ...)
}
\arguments{
\item{path}{path of table file, must be a \code{'csv'} file}

\item{subject}{'RAVE' subject ID or instance}

\item{use_fs}{whether to use 'FreeSurfer' files to calculate other
coordinates}

\item{dry_run}{whether to dry-run the process; if true, then the table
will be generated but not saved to subject's meta folder}

\item{...}{passed to \code{\link[utils]{read.csv}}}
}
\value{
Nothing, the electrode information will be written directly to the
subject's meta directory
}
\description{
Import electrode table into subject meta folder
}
