% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rave-snippets.R
\name{rave-snippet}
\alias{rave-snippet}
\alias{update_local_snippet}
\alias{load_snippet}
\title{'RAVE' code snippets}
\usage{
update_local_snippet(force = TRUE)

load_snippet(topic, local = TRUE)
}
\arguments{
\item{force}{whether to force updating the snippets; default is true}

\item{topic}{snippet topic}

\item{local}{whether to use local snippets first before requesting online
repository}
}
\value{
`load_snippet` returns snippet as a function, others return nothing
}
\description{
Run snippet code
}
\examples{

if(!is_on_cran()) {

  update_local_snippet()
  snippet <- load_snippet("dummy-snippet")

  # Read snippet documentation
  print(snippet)

  # Run snippet as a function
  snippet("this is an input")
}

}
