% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{rave_command_line_path}
\alias{rave_command_line_path}
\alias{normalize_commandline_path}
\alias{cmd_dcm2niix}
\alias{cmd_freesurfer_home}
\alias{cmd_fsl_home}
\alias{cmd_homebrew}
\alias{is_dry_run}
\title{Find and execute external command-line tools}
\usage{
normalize_commandline_path(
  path,
  type = c("dcm2niix", "freesurfer", "fsl", "others"),
  unset = NA
)

cmd_dcm2niix(error_on_missing = TRUE, unset = NA)

cmd_freesurfer_home(error_on_missing = TRUE, unset = NA)

cmd_fsl_home(error_on_missing = TRUE, unset = NA)

cmd_homebrew(error_on_missing = TRUE, unset = NA)

is_dry_run()
}
\arguments{
\item{path}{path to normalize}

\item{type}{type of command}

\item{unset}{default to return if the command is not found}

\item{error_on_missing}{whether to raise errors if command is missing}
}
\value{
Normalized path to the command, or \code{unset} if command is
missing.
}
\description{
Find and execute external command-line tools
}
