% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{raveio-constants}
\alias{raveio-constants}
\alias{SIGNAL_TYPES}
\alias{LOCATION_TYPES}
\title{The constant variables}
\format{
An object of class \code{character} of length 6.

An object of class \code{character} of length 5.
}
\usage{
SIGNAL_TYPES

LOCATION_TYPES
}
\description{
The constant variables
}
\details{
\code{SIGNAL_TYPES} has the following options: \code{'LFP'}, \code{'Spike'},
\code{'EKG'}, \code{'Audio'}, \code{'Photodiode'}, or \code{'Unknown'}. As
of 'raveio' \code{0.0.6}, only \code{'LFP'} (see \code{\link{LFP_electrode}})
signal type is supported.


\code{LOCATION_TYPES} is a list of the electrode location types:
\code{'iEEG'} (this includes the next two), \code{'sEEG'} (stereo),
\code{'ECoG'} (surface), \code{'EEG'} (scalp),
\code{'Others'}. See field \code{'location'} in
\code{\link{RAVEAbstarctElectrode}}
}
\keyword{datasets}
