% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{read-write-fst}
\alias{read-write-fst}
\alias{write_fst}
\alias{read_fst}
\title{Read a 'fst' file}
\usage{
write_fst(x, path, ...)

read_fst(path, ..., as.data.table = TRUE)
}
\arguments{
\item{x}{data frame to write to path}

\item{path}{path to 'fst' file: must not be connection.}

\item{...}{passed to \code{\link[fst]{read_fst}} or \code{\link[fst]{write_fst}}}

\item{as.data.table}{passed to \code{\link[fst]{read_fst}}}
}
\description{
Read a 'fst' file
}
