% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab.R
\name{read_mat}
\alias{read_mat}
\title{Read 'Matlab' Files}
\usage{
read_mat(file)
}
\arguments{
\item{file}{path to a 'Matlab' file}
}
\value{
A list of All the data stored in the file
}
\description{
A compatible reader that can read both 'Matlab'
files prior and after version 6.0
}
\details{
\code{\link[R.matlab]{readMat}} is used to read files
prior to version 6. After version 6, 'Matlab' uses 'HDF5' format
to store its data, hence \code{\link[hdf5r]{H5File}} is used to
read the file.

The performance of \code{read_mat} can be limited when
the file is too big or has many datasets as it reads all the
data contained in 'Matlab' file into memory.
}
