% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/reexport-ieegio.R,
%   R/reexport-ravepipeline.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{glue}
\alias{\%...>\%}
\alias{\%...T>\%}
\alias{\%...!\%}
\alias{\%...T!\%}
\alias{LazyH5}
\alias{LazyFST}
\alias{convert_fst_to_hdf5}
\alias{convert_fst_to_csv}
\alias{raveio_setopt}
\alias{raveio_getopt}
\alias{raveio_resetopt}
\alias{raveio_confpath}
\alias{PipelineTools}
\alias{PipelineResult}
\alias{pipeline}
\alias{pipeline_from_path}
\alias{pipeline_install_local}
\alias{pipeline_install_github}
\alias{pipeline_root}
\alias{pipeline_list}
\alias{pipeline_find}
\alias{pipeline_attach}
\alias{pipeline_run}
\alias{pipeline_clean}
\alias{pipeline_run_bare}
\alias{pipeline_eval}
\alias{pipeline_render}
\alias{configure_knitr}
\alias{pipeline_setup_rmd}
\alias{load_targets}
\alias{pipeline_target_names}
\alias{pipeline_debug}
\alias{pipeline_dep_targets}
\alias{pipeline_visualize}
\alias{pipeline_progress}
\alias{pipeline_fork}
\alias{pipeline_build}
\alias{pipeline_read}
\alias{pipeline_vartable}
\alias{pipeline_hasname}
\alias{pipeline_watch}
\alias{pipeline_create_template}
\alias{pipeline_create_subject_pipeline}
\alias{pipeline_description}
\alias{pipeline_settings_set}
\alias{pipeline_settings_get}
\alias{pipeline_load_extdata}
\alias{pipeline_save_extdata}
\alias{pipeline_shared}
\alias{pipeline_set_preferences}
\alias{pipeline_get_preferences}
\alias{pipeline_has_preferences}
\alias{pipeline_collection}
\alias{PipelineCollections}
\alias{load_snippet}
\alias{update_local_snippet}
\alias{install_snippet}
\alias{list_snippets}
\alias{module_add}
\alias{install_modules}
\alias{module_registry}
\alias{module_registry2}
\alias{get_modules_registries}
\alias{get_module_description}
\alias{add_module_registry}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{glue}{\code{\link[glue]{glue}}}

  \item{ieegio}{\code{\link[ieegio:convert-fst]{convert_fst_to_csv}}, \code{\link[ieegio:convert-fst]{convert_fst_to_hdf5}}, \code{\link[ieegio]{LazyFST}}, \code{\link[ieegio]{LazyH5}}}

  \item{promises}{\code{\link[promises:pipes]{\%...!\%}}, \code{\link[promises:pipes]{\%...>\%}}, \code{\link[promises:pipes]{\%...T!\%}}, \code{\link[promises:pipes]{\%...T>\%}}}

  \item{ravepipeline}{\code{\link[ravepipeline:module_registry]{add_module_registry}}, \code{\link[ravepipeline:pipeline-knitr-markdown]{configure_knitr}}, \code{\link[ravepipeline:module_registry]{get_module_description}}, \code{\link[ravepipeline:module_registry]{get_modules_registries}}, \code{\link[ravepipeline]{install_modules}}, \code{\link[ravepipeline:rave-snippet]{install_snippet}}, \code{\link[ravepipeline:rave-snippet]{list_snippets}}, \code{\link[ravepipeline:rave-snippet]{load_snippet}}, \code{\link[ravepipeline:rave-pipeline]{load_targets}}, \code{\link[ravepipeline]{module_add}}, \code{\link[ravepipeline]{module_registry}}, \code{\link[ravepipeline:module_registry]{module_registry2}}, \code{\link[ravepipeline]{pipeline}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_attach}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_build}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_clean}}, \code{\link[ravepipeline]{pipeline_collection}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_create_subject_pipeline}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_create_template}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_debug}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_dep_targets}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_description}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_eval}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_find}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_fork}}, \code{\link[ravepipeline:pipeline]{pipeline_from_path}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_get_preferences}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_has_preferences}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_hasname}}, \code{\link[ravepipeline:pipeline_install]{pipeline_install_github}}, \code{\link[ravepipeline:pipeline_install]{pipeline_install_local}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_list}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_load_extdata}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_progress}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_read}}, \code{\link[ravepipeline:pipeline-knitr-markdown]{pipeline_render}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_root}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_run}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_run_bare}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_save_extdata}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_set_preferences}}, \code{\link[ravepipeline:pipeline_settings_get_set]{pipeline_settings_get}}, \code{\link[ravepipeline:pipeline_settings_get_set]{pipeline_settings_set}}, \code{\link[ravepipeline:pipeline-knitr-markdown]{pipeline_setup_rmd}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_shared}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_target_names}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_vartable}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_visualize}}, \code{\link[ravepipeline:rave-pipeline]{pipeline_watch}}, \code{\link[ravepipeline]{PipelineCollections}}, \code{\link[ravepipeline]{PipelineResult}}, \code{\link[ravepipeline]{PipelineTools}}, \code{\link[ravepipeline:raveio-option]{raveio_confpath}}, \code{\link[ravepipeline:raveio-option]{raveio_getopt}}, \code{\link[ravepipeline:raveio-option]{raveio_resetopt}}, \code{\link[ravepipeline:raveio-option]{raveio_setopt}}, \code{\link[ravepipeline:rave-snippet]{update_local_snippet}}}
}}

