% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{glue}
\alias{\%...>\%}
\alias{\%...T>\%}
\alias{\%...!\%}
\alias{\%...T!\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{glue}{\code{\link[glue]{glue}}}

  \item{promises}{\code{\link[promises:pipes]{\%...!\%}}, \code{\link[promises:pipes]{\%...>\%}}, \code{\link[promises:pipes]{\%...T!\%}}, \code{\link[promises:pipes]{\%...T>\%}}}
}}

