% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{safe_read_csv}
\alias{safe_read_csv}
\title{Read csv Files with Given Column Classes}
\usage{
safe_read_csv(
  file,
  header = TRUE,
  sep = ",",
  colClasses = NA,
  skip = 0,
  quote = "\\"",
  ...,
  stringsAsFactors = FALSE
)
}
\arguments{
\item{file, header, sep, colClasses, skip, quote, stringsAsFactors, ...}{passed to \code{\link[utils]{read.csv}}}
}
\value{
A data frame
}
\description{
Read csv Files with Given Column Classes
}
\details{
Reading a csv file using builtin function
\code{\link[utils]{read.csv}} might result in some unexpected
behavior. \code{safe_read_csv} does some preprocessing on the
format so that it take cares of the following cases.

1. If \code{skip} exceeds the maximum rows of the data, return
a blank data frame instead of raising error.

2. If row names are included in the file, \code{colClasses}
automatically skip that column and starts from the second column

3. If length of \code{colClasses} does not equal to the number of
columns, instead of cycling the class types, we set those columns
to be \code{NA} type and let \code{\link[utils]{read.csv}} decide
the default types.

4. \code{stringsAsFactors} is by default \code{FALSE} to be
consistent with R 4.0, if the function is called in R 3.x.
}
