% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{safe_write_csv}
\alias{safe_write_csv}
\title{Save Data to csv Files but Rename Old File Instead of Overwrite}
\usage{
safe_write_csv(x, file, ..., quiet = FALSE)
}
\arguments{
\item{x, file}{see also \code{\link[utils]{write.csv}}}

\item{...}{pass to \code{\link[utils]{write.csv}}}

\item{quiet}{whether to suppress overwrite message}
}
\value{
Normalized path of \code{file}
}
\description{
Save Data to csv Files but Rename Old File Instead of Overwrite
}
