% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{test_hdspeed}
\alias{test_hdspeed}
\title{Simple Hard Disk Speed Test}
\usage{
test_hdspeed(path, file_size = 1e+07, quiet = FALSE)
}
\arguments{
\item{path}{an existing directory where to test speed}

\item{file_size}{in bytes, default is 10 MB}

\item{quiet}{should verbose messages be suppressed?}
}
\value{
A vector of two: writing and reading speed in MB per seconds.
}
\description{
Simple Hard Disk Speed Test
}
